+++ app/src/main/java/org/fairscan/app/ui/screens/crop/MagnifyingGlass.kt
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License along with
  * this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.fairscan.app.ui.screens.edit
+package org.fairscan.app.ui.screens.crop
 
 import android.graphics.Bitmap
 import androidx.compose.foundation.Canvas
