+++ app/src/test/java/org/fairscan/app/ui/screens/crop/MagnifyingGlassTest.kt
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License along with
  * this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.fairscan.app.ui.screens.edit
+package org.fairscan.app.ui.screens.crop
 
 import androidx.compose.ui.geometry.Offset
 import org.assertj.core.api.Assertions.assertThat
