+++ app/src/test/java/org/fairscan/app/ui/screens/crop/QuadCoordinateUtilsTest.kt
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License along with
  * this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.fairscan.app.ui.screens.edit
+package org.fairscan.app.ui.screens.crop
 
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.unit.IntSize
